+++ docs/Changelog.md
@@ -2,6 +2,8 @@
 
 ## Next version
 
+- Color filter theme resources [#1622](https://github.com/mapsforge/mapsforge/pull/1622)
+  - `ThemeCallback.getBitmap`
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/mapsforge/issues?q=is%3Aclosed+milestone%3A0.24.0)